+++ crates/weaver_checker/src/violation.rs
@@ -43,13 +43,15 @@ impl Display for Violation {
             }
             Violation::Advice(Advice {
                 advice_type: r#type,
-                value,
+                advice_context,
                 message,
                 advice_level,
+                signal_type,
+                signal_name,
             }) => {
                 write!(
                     f,
-                    "type={type}, value={value}, message={message}, advice_level={advice_level:?}"
+                    "type={type}, context={advice_context}, message={message}, advice_level={advice_level:?}, signal_type={signal_type:?}, signal_name={signal_name:?}"
                 )
             }
         }
@@ -87,12 +89,27 @@ pub enum AdviceLevel {
 /// Represents a live check advice
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize, JsonSchema)]
 pub struct Advice {
-    /// The type of advice e.g. "is_deprecated"
+    /// The type of advice e.g. "is_deprecated". This should be a short,
+    /// machine-readable string that categorizes the advice.
     pub advice_type: String,
-    /// The value of the advice e.g. "true"
-    pub value: Value,
-    /// The message of the advice e.g. "This attribute is deprecated"
+
+    /// The context associated with the advice e.g. { "attribute_name": "foo.bar", "attribute_value": "bar" }
+    /// The context should contain all dynamic parts of the message
+    /// Context values may be used with custom templates and filters to customize reports.
+    pub advice_context: Value,
+
+    /// The human-readable message of the advice e.g. "This attribute 'foo.bar' is deprecated, reason: 'use foo.baz'"
+    /// The message, along with signal_name and signal_type, should contain enough information to understand the advice and
+    /// identify the issue and how to fix it.
+    /// Some of the values used in the message may be also present in the `advice_context` field to support report customization.
     pub message: String,
+
     /// The level of the advice e.g. "violation"
     pub advice_level: AdviceLevel,
+
+    /// The signal type the advice applies to: "span", "metric", "entity", "log" (aka "event"), or "profile"
+    pub signal_type: Option<String>,
+
+    /// The signal name the advice applies to e.g. "http.server.request.duration".
+    pub signal_name: Option<String>,
 }