+++ crates/weaver_live_check/src/advice.rs
@@ -9,7 +9,7 @@ use std::{
 };
 
 use serde::Serialize;
-use serde_json::Value;
+use serde_json::json;
 use weaver_checker::{
     violation::{Advice, AdviceLevel, Violation},
     Engine,
@@ -27,7 +27,12 @@ use weaver_semconv::{
 
 use crate::{
     live_checker::LiveChecker, sample_attribute::SampleAttribute, sample_metric::SampleInstrument,
-    Error, SampleRef,
+    Error, Sample, SampleRef, ATTRIBUTE_NAME_ADVICE_CONTEXT_KEY, ATTRIBUTE_TYPE_ADVICE_CONTEXT_KEY,
+    ATTRIBUTE_VALUE_ADVICE_CONTEXT_KEY, DEPRECATED_ADVICE_TYPE,
+    DEPRECATION_NOTE_ADVICE_CONTEXT_KEY, DEPRECATION_REASON_ADVICE_CONTEXT_KEY,
+    EXPECTED_VALUE_ADVICE_CONTEXT_KEY, INSTRUMENT_ADVICE_CONTEXT_KEY, NOT_STABLE_ADVICE_TYPE,
+    STABILITY_ADVICE_CONTEXT_KEY, TYPE_MISMATCH_ADVICE_TYPE, UNDEFINED_ENUM_VARIANT_ADVICE_TYPE,
+    UNEXPECTED_INSTRUMENT_ADVICE_TYPE, UNIT_ADVICE_CONTEXT_KEY, UNIT_MISMATCH_ADVICE_TYPE,
 };
 
 /// Embedded default live check rego policies
@@ -47,17 +52,18 @@ pub trait Advisor {
     fn advise(
         &mut self,
         sample: SampleRef<'_>,
+        signal: &Sample,
         registry_attribute: Option<Rc<Attribute>>,
         registry_group: Option<Rc<ResolvedGroup>>,
     ) -> Result<Vec<Advice>, Error>;
 }
 
-fn deprecated_to_value(deprecated: &Deprecated) -> Value {
+fn deprecated_to_reason(deprecated: &Deprecated) -> String {
     match deprecated {
-        Deprecated::Renamed { .. } => Value::String("renamed".to_owned()),
-        Deprecated::Obsoleted { .. } => Value::String("obsoleted".to_owned()),
+        Deprecated::Renamed { .. } => "renamed".to_owned(),
+        Deprecated::Obsoleted { .. } => "obsoleted".to_owned(),
         Deprecated::Uncategorized { .. } | Deprecated::Unspecified { .. } => {
-            Value::String("uncategorized".to_owned())
+            "uncategorized".to_owned()
         }
     }
 }
@@ -68,33 +74,54 @@ impl Advisor for DeprecatedAdvisor {
     fn advise(
         &mut self,
         sample: SampleRef<'_>,
+        signal: &Sample,
         registry_attribute: Option<Rc<Attribute>>,
         registry_group: Option<Rc<ResolvedGroup>>,
     ) -> Result<Vec<Advice>, Error> {
         match sample {
-            SampleRef::Attribute(_sample_attribute) => {
+            SampleRef::Attribute(sample_attribute) => {
                 let mut advices = Vec::new();
                 if let Some(attribute) = registry_attribute {
                     if let Some(deprecated) = &attribute.deprecated {
                         advices.push(Advice {
-                            advice_type: "deprecated".to_owned(),
-                            value: deprecated_to_value(deprecated),
-                            message: deprecated.to_string(),
+                            advice_type: DEPRECATED_ADVICE_TYPE.to_owned(),
+                            advice_context: json!({
+                                ATTRIBUTE_NAME_ADVICE_CONTEXT_KEY: sample_attribute.name.clone(),
+                                DEPRECATION_REASON_ADVICE_CONTEXT_KEY: deprecated_to_reason(deprecated),
+                                DEPRECATION_NOTE_ADVICE_CONTEXT_KEY: deprecated.to_string(),
+                            }),
+                            message: format!(
+                                "Attribute '{}' is deprecated; reason = '{}', note = '{}'.",
+                                sample_attribute.name.clone(),
+                                deprecated_to_reason(deprecated),
+                                deprecated
+                            ),
                             advice_level: AdviceLevel::Violation,
+                            signal_type: signal.signal_type(),
+                            signal_name: signal.signal_name(),
                         });
                     }
                 }
                 Ok(advices)
             }
-            SampleRef::Metric(_sample_metric) => {
+            SampleRef::Metric(sample_metric) => {
                 let mut advices = Vec::new();
                 if let Some(group) = registry_group {
                     if let Some(deprecated) = &group.deprecated {
                         advices.push(Advice {
-                            advice_type: "deprecated".to_owned(),
-                            value: deprecated_to_value(deprecated),
-                            message: deprecated.to_string(),
+                            advice_type: DEPRECATED_ADVICE_TYPE.to_owned(),
+                            advice_context: json!({
+                                DEPRECATION_REASON_ADVICE_CONTEXT_KEY: deprecated_to_reason(deprecated),
+                                DEPRECATION_NOTE_ADVICE_CONTEXT_KEY: deprecated,
+                            }),
+                            message: format!(
+                                "Metric is deprecated; reason = {}, note = {}",
+                                deprecated_to_reason(deprecated),
+                                deprecated
+                            ),
                             advice_level: AdviceLevel::Violation,
+                            signal_type: Some("metric".to_owned()),
+                            signal_name: Some(sample_metric.name.clone()),
                         });
                     }
                 }
@@ -114,20 +141,30 @@ impl Advisor for StabilityAdvisor {
     fn advise(
         &mut self,
         sample: SampleRef<'_>,
+        parent_signal: &Sample,
         registry_attribute: Option<Rc<Attribute>>,
         registry_group: Option<Rc<ResolvedGroup>>,
     ) -> Result<Vec<Advice>, Error> {
         match sample {
-            SampleRef::Attribute(_sample_attribute) => {
+            SampleRef::Attribute(sample_attribute) => {
                 let mut advices = Vec::new();
                 if let Some(attribute) = registry_attribute {
                     match attribute.stability {
                         Some(ref stability) if *stability != Stability::Stable => {
                             advices.push(Advice {
-                                advice_type: "stability".to_owned(),
-                                value: Value::String(stability.to_string()),
-                                message: "Is not stable".to_owned(),
+                                advice_type: NOT_STABLE_ADVICE_TYPE.to_owned(),
+                                advice_context: json!({
+                                    ATTRIBUTE_NAME_ADVICE_CONTEXT_KEY: sample_attribute.name.clone(),
+                                    STABILITY_ADVICE_CONTEXT_KEY: stability,
+                                }),
+                                message: format!(
+                                    "Attribute '{}' is not stable; stability = {}.",
+                                    sample_attribute.name.clone(),
+                                    stability
+                                ),
                                 advice_level: AdviceLevel::Improvement,
+                                signal_type: parent_signal.signal_type(),
+                                signal_name: parent_signal.signal_name(),
                             });
                         }
                         _ => {}
@@ -141,10 +178,17 @@ impl Advisor for StabilityAdvisor {
                     match group.stability {
                         Some(ref stability) if *stability != Stability::Stable => {
                             advices.push(Advice {
-                                advice_type: "stability".to_owned(),
-                                value: Value::String(stability.to_string()),
-                                message: "Is not stable".to_owned(),
+                                advice_type: NOT_STABLE_ADVICE_TYPE.to_owned(),
+                                advice_context: json!({
+                                    STABILITY_ADVICE_CONTEXT_KEY: stability,
+                                }),
+                                message: format!(
+                                    "Metric is not stable; stability = {}.",
+                                    stability
+                                ),
                                 advice_level: AdviceLevel::Improvement,
+                                signal_type: parent_signal.signal_type(),
+                                signal_name: parent_signal.signal_name(),
                             });
                         }
                         _ => {}
@@ -175,6 +219,7 @@ pub struct TypeAdvisor;
 fn check_attributes(
     semconv_attributes: &[Attribute],
     sample_attributes: &[SampleAttribute],
+    sample: &Sample,
 ) -> Vec<Advice> {
     // Create a HashSet of attribute names for O(1) lookups
     let attribute_set: HashSet<_> = sample_attributes.iter().map(|attr| &attr.name).collect();
@@ -186,32 +231,47 @@ fn check_attributes(
                 RequirementLevel::Basic(BasicRequirementLevelSpec::Required) => (
                     "required_attribute_not_present".to_owned(),
                     AdviceLevel::Violation,
-                    "Required attribute is not present".to_owned(),
+                    format!(
+                        "Required attribute '{}' is not present.",
+                        semconv_attribute.name
+                    ),
                 ),
                 RequirementLevel::Basic(BasicRequirementLevelSpec::Recommended)
                 | RequirementLevel::Recommended { .. } => (
                     "recommended_attribute_not_present".to_owned(),
                     AdviceLevel::Improvement,
-                    "Recommended attribute is not present".to_owned(),
+                    format!(
+                        "Recommended attribute '{}' is not present.",
+                        semconv_attribute.name
+                    ),
                 ),
                 RequirementLevel::Basic(BasicRequirementLevelSpec::OptIn)
                 | RequirementLevel::OptIn { .. } => (
                     "opt_in_attribute_not_present".to_owned(),
                     AdviceLevel::Information,
-                    "Opt-in attribute is not present".to_owned(),
+                    format!(
+                        "Opt-in attribute '{}' is not present.",
+                        semconv_attribute.name
+                    ),
                 ),
                 RequirementLevel::ConditionallyRequired { .. } => (
                     "conditionally_required_attribute_not_present".to_owned(),
                     AdviceLevel::Information,
-                    "Conditionally required attribute is not present".to_owned(),
+                    format!(
+                        "Conditionally required attribute '{}' is not present.",
+                        semconv_attribute.name
+                    ),
                 ),
             };
-
             advice_list.push(Advice {
                 advice_type,
-                value: Value::String(semconv_attribute.name.clone()),
+                advice_context: json!({
+                    ATTRIBUTE_NAME_ADVICE_CONTEXT_KEY: semconv_attribute.name.clone()
+                }),
                 message,
                 advice_level,
+                signal_type: sample.signal_type(),
+                signal_name: sample.signal_name(),
             });
         }
     }
@@ -222,6 +282,7 @@ impl Advisor for TypeAdvisor {
     fn advise(
         &mut self,
         sample: SampleRef<'_>,
+        parent_signal: &Sample,
         registry_attribute: Option<Rc<Attribute>>,
         registry_group: Option<Rc<ResolvedGroup>>,
     ) -> Result<Vec<Advice>, Error> {
@@ -255,10 +316,15 @@ impl Advisor for TypeAdvisor {
                                     && attribute_type != &PrimitiveOrArrayTypeSpec::Int
                                 {
                                     return Ok(vec![Advice {
-                                        advice_type: "type_mismatch".to_owned(),
-                                        value: Value::String(attribute_type.to_string()),
-                                        message: "Type should be `string` or `int`".to_owned(),
+                                        advice_type: TYPE_MISMATCH_ADVICE_TYPE.to_owned(),
+                                        advice_context: json!({
+                                            ATTRIBUTE_NAME_ADVICE_CONTEXT_KEY: sample_attribute.name.clone(),
+                                            ATTRIBUTE_TYPE_ADVICE_CONTEXT_KEY: attribute_type,
+                                        }),
+                                        message: format!("Enum attribute '{}' has type '{}'. Enum value type should be 'string' or 'int'.", sample_attribute.name, attribute_type),
                                         advice_level: AdviceLevel::Violation,
+                                        signal_type: parent_signal.signal_type(),
+                                        signal_name: parent_signal.signal_name(),
                                     }]);
                                 } else {
                                     return Ok(Vec::new());
@@ -268,10 +334,19 @@ impl Advisor for TypeAdvisor {
 
                         if !attribute_type.is_compatible(semconv_attribute_type) {
                             Ok(vec![Advice {
-                                advice_type: "type_mismatch".to_owned(),
-                                value: Value::String(attribute_type.to_string()),
-                                message: format!("Type should be `{semconv_attribute_type}`"),
+                                advice_type: TYPE_MISMATCH_ADVICE_TYPE.to_owned(),
+                                advice_context: json!({
+                                    ATTRIBUTE_NAME_ADVICE_CONTEXT_KEY: sample_attribute.name.clone(),
+                                    ATTRIBUTE_TYPE_ADVICE_CONTEXT_KEY: attribute_type,
+                                    EXPECTED_VALUE_ADVICE_CONTEXT_KEY: semconv_attribute_type,
+                                }),
+                                message: format!(
+                                    "Attribute '{}' has type '{}'. Type should be '{}'.",
+                                    sample_attribute.name, attribute_type, semconv_attribute_type
+                                ),
                                 advice_level: AdviceLevel::Violation,
+                                signal_type: parent_signal.signal_type(),
+                                signal_name: parent_signal.signal_name(),
                             }])
                         } else {
                             Ok(Vec::new())
@@ -283,26 +358,36 @@ impl Advisor for TypeAdvisor {
             SampleRef::Metric(sample_metric) => {
                 // Check the instrument and unit of the metric
                 let mut advice_list = Vec::new();
+
                 if let Some(semconv_metric) = registry_group {
                     match &sample_metric.instrument {
                         SampleInstrument::Unsupported(name) => {
                             advice_list.push(Advice {
-                                advice_type: "unsupported_instrument".to_owned(),
-                                value: name.as_str().into(),
-                                message: "Instrument is not supported".to_owned(),
+                                advice_type: UNEXPECTED_INSTRUMENT_ADVICE_TYPE.to_owned(),
+                                advice_context: json!({
+                                    INSTRUMENT_ADVICE_CONTEXT_KEY: name.clone()
+                                }),
+                                message: format!("Instrument '{name}' is not supported"),
                                 advice_level: AdviceLevel::Violation,
+                                signal_type: parent_signal.signal_type(),
+                                signal_name: parent_signal.signal_name(),
                             });
                         }
                         SampleInstrument::Supported(sample_instrument) => {
                             if let Some(semconv_instrument) = &semconv_metric.instrument {
                                 if semconv_instrument != sample_instrument {
                                     advice_list.push(Advice {
-                                        advice_type: "instrument_mismatch".to_owned(),
-                                        value: Value::String(sample_instrument.to_string()),
+                                        advice_type: UNEXPECTED_INSTRUMENT_ADVICE_TYPE.to_owned(),
+                                        advice_context: json!({
+                                            INSTRUMENT_ADVICE_CONTEXT_KEY: sample_instrument,
+                                            EXPECTED_VALUE_ADVICE_CONTEXT_KEY: semconv_instrument,
+                                        }),
                                         message: format!(
-                                            "Instrument should be `{semconv_instrument}`"
+                                            "Instrument should be '{semconv_instrument}', but found '{sample_instrument}'."
                                         ),
                                         advice_level: AdviceLevel::Violation,
+                                        signal_type: parent_signal.signal_type(),
+                                        signal_name: parent_signal.signal_name(),
                                     });
                                 }
                             }
@@ -312,10 +397,18 @@ impl Advisor for TypeAdvisor {
                     if let Some(semconv_unit) = &semconv_metric.unit {
                         if semconv_unit != &sample_metric.unit {
                             advice_list.push(Advice {
-                                advice_type: "unit_mismatch".to_owned(),
-                                value: Value::String(sample_metric.unit.clone()),
-                                message: format!("Unit should be `{semconv_unit}`"),
+                                advice_type: UNIT_MISMATCH_ADVICE_TYPE.to_owned(),
+                                advice_context: json!({
+                                    UNIT_ADVICE_CONTEXT_KEY: sample_metric.unit.clone(),
+                                    EXPECTED_VALUE_ADVICE_CONTEXT_KEY: semconv_unit.clone(),
+                                }),
+                                message: format!(
+                                    "Unit should be '{semconv_unit}', but found '{}'.",
+                                    sample_metric.unit
+                                ),
                                 advice_level: AdviceLevel::Violation,
+                                signal_type: parent_signal.signal_type(),
+                                signal_name: parent_signal.signal_name(),
                             });
                         }
                     }
@@ -327,6 +420,7 @@ impl Advisor for TypeAdvisor {
                     let advice_list = check_attributes(
                         &semconv_metric.attributes,
                         &sample_number_data_point.attributes,
+                        parent_signal,
                     );
 
                     Ok(advice_list)
@@ -339,6 +433,7 @@ impl Advisor for TypeAdvisor {
                     Ok(check_attributes(
                         &semconv_metric.attributes,
                         &sample_histogram_data_point.attributes,
+                        parent_signal,
                     ))
                 } else {
                     Ok(Vec::new())
@@ -355,6 +450,7 @@ impl Advisor for EnumAdvisor {
     fn advise(
         &mut self,
         sample: SampleRef<'_>,
+        signal: &Sample,
         registry_attribute: Option<Rc<Attribute>>,
         _registry_group: Option<Rc<ResolvedGroup>>,
     ) -> Result<Vec<Advice>, Error> {
@@ -398,10 +494,15 @@ impl Advisor for EnumAdvisor {
 
                             if !is_found {
                                 return Ok(vec![Advice {
-                                    advice_type: "undefined_enum_variant".to_owned(),
-                                    value: attribute_value.clone(),
-                                    message: "Is not a defined variant".to_owned(),
+                                    advice_type: UNDEFINED_ENUM_VARIANT_ADVICE_TYPE.to_owned(),
+                                    advice_context: json!({
+                                        ATTRIBUTE_NAME_ADVICE_CONTEXT_KEY: sample_attribute.name.clone(),
+                                        ATTRIBUTE_VALUE_ADVICE_CONTEXT_KEY: attribute_value,
+                                    }),
+                                    message: format!("Enum attribute '{}' has value '{}' which is not documented.", sample_attribute.name, attribute_value.as_str().unwrap_or("")),
                                     advice_level: AdviceLevel::Information,
+                                    signal_type: signal.signal_type(),
+                                    signal_name: signal.signal_name(),
                                 }]);
                             }
                         }
@@ -512,6 +613,7 @@ impl Advisor for RegoAdvisor {
     fn advise(
         &mut self,
         sample: SampleRef<'_>,
+        _signal: &Sample,
         registry_attribute: Option<Rc<Attribute>>,
         registry_group: Option<Rc<ResolvedGroup>>,
     ) -> Result<Vec<Advice>, Error> {
@@ -525,6 +627,10 @@ impl Advisor for RegoAdvisor {
 
 #[cfg(test)]
 mod tests {
+    use std::collections::HashMap;
+
+    use crate::sample_metric::SampleMetric;
+
     use super::*;
     use weaver_resolved_schema::attribute::Attribute;
     use weaver_semconv::attribute::{
@@ -598,11 +704,19 @@ mod tests {
         // Provide no attributes
         let sample_attributes = vec![];
 
-        let advice = check_attributes(&semconv_attributes, &sample_attributes);
+        // Use a dummy Sample for signal_type and signal_name
+        let sample = Sample::Metric(SampleMetric {
+            name: "test_metric".to_owned(),
+            unit: "".to_owned(),
+            data_points: None,
+            instrument: SampleInstrument::Supported(weaver_semconv::group::InstrumentSpec::Counter),
+            live_check_result: None,
+        });
+        let advice = check_attributes(&semconv_attributes, &sample_attributes, &sample);
         assert_eq!(advice.len(), 6);
 
         // Verify each advice type and level
-        let advice_map: std::collections::HashMap<_, _> = advice
+        let advice_map: HashMap<_, _> = advice
             .iter()
             .map(|a| (a.advice_type.clone(), a.advice_level.clone()))
             .collect();
@@ -660,7 +774,15 @@ mod tests {
             create_sample_attribute("attr2"),
         ];
 
-        let advice = check_attributes(&semconv_attributes, &sample_attributes);
+        // Use a dummy Sample for signal_type and signal_name
+        let sample = Sample::Metric(SampleMetric {
+            name: "test_metric".to_owned(),
+            unit: "".to_owned(),
+            data_points: None,
+            instrument: SampleInstrument::Supported(weaver_semconv::group::InstrumentSpec::Counter),
+            live_check_result: None,
+        });
+        let advice = check_attributes(&semconv_attributes, &sample_attributes, &sample);
         assert!(advice.is_empty());
     }
 }